update rag & full-context function
